xilinx_platform: rename header file
Use a more suggestive header file name for xilinx platform driver.
